+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Fix parsing values provided with abbreviated long option names
 * Fix any object type to symbol conversion failure to return an undefined value
 * Fix parameter validation to stop transforming hash value into an array
+* Fix help display when formatting only short options
 
 ## [v0.2.0] - 2021-08-04